User Authentication
Why?
GIVEN that I'm a user and I have signed up previously
WHEN I come to signin page
AND I fill the form correctly
I SHOULD be logged in
ALSO WHEN I click on log out
I SHOULD be logged out

How?
Added methods to the Session Controller to process form login, and also modified view pages to display Log out link when user is Logged in, and Sign in link when user is logged out.

https://www.pivotaltracker.com/story/show/112318227

[finishes #112318227]