Manage past bookings
Why?
Given that I am a user and I click on past bookings
I should see a list of all my previous bookings and be able to delete a booking
if the flight has not departed

How?
Use index action to display all bookings by Id and used helper methods to check if the date is past before displaying delete button

https://www.pivotaltracker.com/story/show/112728391

[finishes #112728391]