Manage User Booking
Why?
GIVEN that I am a registered user
And I sign in
WHEN I click on manage bookings
I should be able to edit my booking details for flights that have not departed

How?
Create action in booking controller to edit booking for flight that is still available

https://www.pivotaltracker.com/story/show/112728597

[finishes #112728597]

finish edit booking